refactor: Move comparators to better package structure
I'm adding a ton of comparators here and some seem like they have a
"right" place to go - for parts of a listing where we have broken off
that "thing" into it's own package, but for other parts of a listing
where we have not done that refactor yet I'm just putting the
comparators in here.

[#OCD-4064]
