+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedproduct/service/CqmResultsService.java
@@ -11,13 +11,13 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import gov.healthit.chpl.certifiedproduct.service.comparator.CQMCriteriaComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.CQMResultComparator;
 import gov.healthit.chpl.dao.CQMResultDAO;
 import gov.healthit.chpl.dao.CQMResultDetailsDAO;
 import gov.healthit.chpl.domain.CQMCriterion;
 import gov.healthit.chpl.domain.CQMResultCertification;
 import gov.healthit.chpl.domain.CQMResultDetails;
+import gov.healthit.chpl.domain.comparator.CQMCriteriaComparator;
+import gov.healthit.chpl.domain.comparator.CQMResultComparator;
 import gov.healthit.chpl.dto.CQMResultDetailsDTO;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.manager.DimensionalDataManager;
