+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedproduct/service/ListingService.java
@@ -13,13 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
-import gov.healthit.chpl.certifiedproduct.service.comparator.CertifiedProductAccessibilityStandardComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.CertifiedProductComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.CertifiedProductQmsStandardComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.CertifiedProductTargetedUserComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.CertifiedProductTestingLabComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.ChplProductNumberHistoryComparator;
-import gov.healthit.chpl.certifiedproduct.service.comparator.DirectReviewComparator;
+import gov.healthit.chpl.compliance.directreview.DirectReviewComparator;
 import gov.healthit.chpl.compliance.directreview.DirectReviewSearchService;
 import gov.healthit.chpl.compliance.surveillance.SurveillanceManager;
 import gov.healthit.chpl.dao.CertifiedProductAccessibilityStandardDAO;
@@ -42,6 +36,12 @@ import gov.healthit.chpl.domain.CertifiedProductTestingLab;
 import gov.healthit.chpl.domain.InheritedCertificationStatus;
 import gov.healthit.chpl.domain.ProductVersion;
 import gov.healthit.chpl.domain.PromotingInteroperabilityUser;
+import gov.healthit.chpl.domain.comparator.CertifiedProductAccessibilityStandardComparator;
+import gov.healthit.chpl.domain.comparator.CertifiedProductComparator;
+import gov.healthit.chpl.domain.comparator.CertifiedProductQmsStandardComparator;
+import gov.healthit.chpl.domain.comparator.CertifiedProductTargetedUserComparator;
+import gov.healthit.chpl.domain.comparator.CertifiedProductTestingLabComparator;
+import gov.healthit.chpl.domain.comparator.ChplProductNumberHistoryComparator;
 import gov.healthit.chpl.domain.compliance.DirectReview;
 import gov.healthit.chpl.dto.CertifiedProductDTO;
 import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
