+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/comparator/CertificationCriterionComparator.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.certifiedproduct.service.comparator;
+package gov.healthit.chpl.domain.comparator;
 
 import java.util.Comparator;
